requirements.txt
asgiref==3.8.1
Django==4.2.21
django-allauth==65.2.0
django-environ==0.12.0
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
